source/mre-plugin-samples/cdk/requirements.txt
aws-cdk.core>=1.85.0,<2.0
aws-cdk.aws-iam>=1.85.0,<2.0
aws-cdk.aws-lambda>=1.85.0,<2.0
boto3<2.0.0
urllib3
requests
requests-aws4auth
